requirements.txt
python-binance>=1.0.19
pandas>=2.2.2
numpy>=1.26.4
matplotlib>=3.9.0
TA-Lib>=0.4.32
python-telegram-bot>=21.0.1
